# schemas/jobs/openapi.yaml
openapi: 3.0.3
info:
  title: Jobs API
  version: "1.0.0"
  description: |
    OpenAPI schema for the Jobs API derived from app/lib/jobTypes.ts.
    This spec includes the Job model and related component schemas (attachments,
    locations, salary, contacts and enums). It intentionally excludes types
    related to the search API (e.g. JobSearchResponse, JobResultItem).

    Notes:
      - id is generated by the system.
      - externalId is provided by the client and is used as the primary key for GET/PUT/PATCH endpoints.
servers:
  - url: https://api.example.com
paths:
  /jobs:
    get:
      summary: Retrieve all jobs (minimal)
      operationId: listJobs
      description: |
        Returns a minimal list of jobs (identifiers, title, approach, and dateClosing).
        Optional query parameters can be used to filter by closing date/time
        and approach. dateClosing filters jobs whose dateClosing is on or after
        the provided date/time.
      parameters:
        - name: dateClosing
          in: query
          required: false
          schema:
            type: string
            format: date-time
          description: Return jobs with dateClosing on or after this date/time (ISO 8601).
        - name: approach
          in: query
          required: false
          schema:
            $ref: "#/components/schemas/Approach"
          description: Filter results to jobs with the specified approach.
      responses:
        "200":
          description: A minimal list of jobs
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/JobSummary"
        "400":
          description: Bad request - invalid query parameter
    post:
      summary: Create a new job
      operationId: createJob
      requestBody:
        description: Job payload to create. id is generated by the system; externalId must be provided by the client.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobCreate"
      responses:
        "201":
          description: Job created
          headers:
            Location:
              description: URL of the newly created job
              schema:
                type: string
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "400":
          description: Bad request - validation error
  /jobs/{externalId}:
    parameters:
      - name: externalId
        in: path
        required: true
        schema:
          type: string
        description: External identifier for the job (used as primary key)
    get:
      summary: Retrieve a job by externalId
      operationId: getJobByExternalId
      responses:
        "200":
          description: Job found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "404":
          description: Job not found
    put:
      summary: Replace an existing job identified by externalId
      operationId: replaceJobByExternalId
      requestBody:
        description: Full job payload to replace an existing job. The resource to replace is identified by the path parameter externalId. externalId in the body, if provided, MUST match the path param[...]
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobCreate"
      responses:
        "200":
          description: Job replaced
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "400":
          description: Bad request - validation error
        "404":
          description: Job not found
    patch:
      summary: Partially update an existing job identified by externalId
      operationId: updateJobByExternalId
      requestBody:
        description: Partial job payload - only include fields to update. externalId is readonly and cannot be modified via PATCH.
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JobUpdate"
      responses:
        "200":
          description: Job updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Job"
        "400":
          description: Bad request - validation error
        "404":
          description: Job not found

components:
  schemas:
    JobSummary:
      type: object
      description: Minimal job listing info used by GET /jobs
      properties:
        id:
          type: string
          description: Internal unique identifier (system-generated)
        externalId:
          type: string
          description: External identifier provided by the client (primary key)
        title:
          type: string
        approach:
          $ref: "#/components/schemas/Approach"
        dateClosing:
          type: string
          format: date-time
          description: Closing date/time for applications
      required:
        - id
        - externalId
        - title
        - approach
        - dateClosing

    Job:
      type: object
      description: Full job posting object
      properties:
        id:
          type: string
          description: Internal unique identifier (generated by the system)
        externalId:
          type: string
          description: External/third-party identifier (provided by client, readonly after creation)
        approach:
          $ref: "#/components/schemas/Approach"
        title:
          type: string
        description:
          type: string
        organisation:
          type: string
        location:
          oneOf:
            - type: array
              description: An array of fixed UK locations (all items are FixedLocations)
              minItems: 1
              items:
                $ref: "#/components/schemas/FixedLocations"
            - type: array
              description: An array of overseas locations (all items are OverseasLocations)
              minItems: 1
              items:
                $ref: "#/components/schemas/OverseasLocations"
        grade:
          anyOf:
            - $ref: "#/components/schemas/Grade"
            - type: string
        assignmentType:
          $ref: "#/components/schemas/Assignments"
        workLocation:
          type: array
          description: Primary work location(s)
          items:
            $ref: "#/components/schemas/WorkLocation"
        workingPattern:
          type: array
          description: Working pattern(s)
          items:
            $ref: "#/components/schemas/WorkingPattern"
        personalSpec:
          type: string
          description: Personal specification content (HTML or text)
        applyDetail:
          type: string
          description: Application details/instructions
        datePosted:
          type: string
          format: date-time
          description: Date the job was posted
        dateClosing:
          type: string
          format: date-time
          description: Closing date/time for applications
        profession:
          $ref: "#/components/schemas/Profession"
        recruitmentEmail:
          type: string
          format: email
        contacts:
          type: array
          items:
            $ref: "#/components/schemas/Contacts"
        nationalityRequirement:
          type: string
        summary:
          type: string
        applyUrl:
          type: string
          format: uri
        benefits:
          type: string
        salary:
          $ref: "#/components/schemas/Salary"
        jobNumbers:
          type: integer
        successProfileDetails:
          type: string
        diversityStatement:
          type: string
        disabilityConfident:
          type: string
        dcStatus:
          $ref: "#/components/schemas/DCStatus"
        redeploymentScheme:
          type: string
        prisonScheme:
          type: string
        veteranScheme:
          type: string
        criminalRecordCheck:
          type: string
        complaintsInfo:
          type: string
        workingForTheCivilService:
          type: string
        eligibilityCheck:
          type: string
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/JobAttachment"
      required:
        - id
        - externalId
        - approach
        - title
        - description
        - organisation
        - location
        - grade
        - assignmentType
        - workLocation
        - workingPattern
        - personalSpec
        - applyDetail
        - datePosted
        - dateClosing
        - profession
        - recruitmentEmail

    JobCreate:
      type: object
      description: Payload for creating a job. id is generated by the system; externalId must be provided by the client.
      properties:
        externalId:
          type: string
          description: External/third-party identifier (provided by client)
        approach:
          $ref: "#/components/schemas/Approach"
        title:
          type: string
        description:
          type: string
        organisation:
          type: string
        location:
          oneOf:
            - type: array
              description: An array of fixed UK locations (all items are FixedLocations)
              minItems: 1
              items:
                $ref: "#/components/schemas/FixedLocations"
            - type: array
              description: An array of overseas locations (all items are OverseasLocations)
              minItems: 1
              items:
                $ref: "#/components/schemas/OverseasLocations"
        grade:
          anyOf:
            - $ref: "#/components/schemas/Grade"
            - type: string
        assignmentType:
          $ref: "#/components/schemas/Assignments"
        workLocation:
          type: array
          items:
            $ref: "#/components/schemas/WorkLocation"
        workingPattern:
          type: array
          items:
            $ref: "#/components/schemas/WorkingPattern"
        personalSpec:
          type: string
        applyDetail:
          type: string
        datePosted:
          type: string
          format: date-time
          description: Date the job was posted
        dateClosing:
          type: string
          format: date-time
        profession:
          $ref: "#/components/schemas/Profession"
        recruitmentEmail:
          type: string
          format: email
        contacts:
          type: array
          items:
            $ref: "#/components/schemas/Contacts"
        nationalityRequirement:
          type: string
        summary:
          type: string
        applyUrl:
          type: string
          format: uri
        benefits:
          type: string
        salary:
          $ref: "#/components/schemas/Salary"
        jobNumbers:
          type: integer
        successProfileDetails:
          type: string
        diversityStatement:
          type: string
        disabilityConfident:
          type: string
        dcStatus:
          $ref: "#/components/schemas/DCStatus"
        redeploymentScheme:
          type: string
        prisonScheme:
          type: string
        veteranScheme:
          type: string
        criminalRecordCheck:
          type: string
        complaintsInfo:
          type: string
        workingForTheCivilService:
          type: string
        eligibilityCheck:
          type: string
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/JobAttachment"
      required:
        - externalId
        - approach
        - title
        - description
        - organisation
        - location
        - grade
        - assignmentType
        - workLocation
        - workingPattern
        - personalSpec
        - applyDetail
        - datePosted
        - dateClosing
        - profession
        - recruitmentEmail

    JobUpdate:
      type: object
      description: Partial update payload for a job. Only include fields you want to change. externalId is readonly and therefore not present here.
      properties:
        approach:
          $ref: "#/components/schemas/Approach"
        title:
          type: string
        description:
          type: string
        organisation:
          type: string
        location:
          oneOf:
            - type: array
              description: An array of fixed UK locations (all items are FixedLocations)
              minItems: 1
              items:
                $ref: "#/components/schemas/FixedLocations"
            - type: array
              description: An array of overseas locations (all items are OverseasLocations)
              minItems: 1
              items:
                $ref: "#/components/schemas/OverseasLocations"
        grade:
          anyOf:
            - $ref: "#/components/schemas/Grade"
            - type: string
        assignmentType:
          $ref: "#/components/schemas/Assignments"
        workLocation:
          type: array
          items:
            $ref: "#/components/schemas/WorkLocation"
        workingPattern:
          type: array
          items:
            $ref: "#/components/schemas/WorkingPattern"
        personalSpec:
          type: string
        applyDetail:
          type: string
        datePosted:
          type: string
          format: date-time
          description: Date the job was posted
        dateClosing:
          type: string
          format: date-time
        profession:
          $ref: "#/components/schemas/Profession"
        recruitmentEmail:
          type: string
          format: email
        contacts:
          type: array
          items:
            $ref: "#/components/schemas/Contacts"
        nationalityRequirement:
          type: string
        summary:
          type: string
        applyUrl:
          type: string
          format: uri
        benefits:
          type: string
        salary:
          $ref: "#/components/schemas/Salary"
        jobNumbers:
          type: integer
        successProfileDetails:
          type: string
        diversityStatement:
          type: string
        disabilityConfident:
          type: string
        dcStatus:
          $ref: "#/components/schemas/DCStatus"
        redeploymentScheme:
          type: string
        prisonScheme:
          type: string
        veteranScheme:
          type: string
        criminalRecordCheck:
          type: string
        complaintsInfo:
          type: string
        workingForTheCivilService:
          type: string
        eligibilityCheck:
          type: string
        attachments:
          type: array
          items:
            $ref: "#/components/schemas/JobAttachment"

    JobAttachment:
      type: object
      properties:
        href:
          type: string
          format: uri
        docName:
          type: string
        docFormat:
          type: string
        fileSize:
          type: string
      required:
        - href
        - docName
        - docFormat

    FixedLocations:
      type: object
      properties:
        uprn:
          type: string
        paoText:
          type: string
          description: Optional — may be omitted or null
        saoText:
          type: string
        streetDescription:
          type: string
        townName:
          type: string
        region:
          type: string
        postcode:
          type: string
        latitude:
          type: number
          format: double
        longitude:
          type: number
          format: double
      required:
        - townName
        - region
        - latitude
        - longitude
      description: >
        Fixed UK address/location fields aligned to the TypeScript interface:
        uprn, paoText, saoText, streetDescription and postcode are optional;
        townName, region, latitude and longitude are required.

    OverseasLocations:
      type: object
      properties:
        countryName:
          type: string
        countryCode:
          type: string
        locationDisplay:
          type: string
      required:
        - countryName
        - countryCode

    Salary:
      type: object
      properties:
        minimum:
          type: number
          format: double
        maximum:
          type: number
          format: double
        currency:
          type: string
        currencySymbol:
          type: string
        salaryDetails:
          type: string
      required:
        - minimum
        - currency

    Contacts:
      type: object
      properties:
        contactName:
          type: string
        contactEmail:
          type: string
          format: email
        contactPhone:
          type: string
      required:
        - contactName
        - contactEmail

    Approach:
      type: string
      description: Recruitment approach
      enum:
        - Internal
        - Across Government
        - External

    Assignments:
      type: string
      description: Assignment / contract type
      enum:
        - Apprentice
        - Fixed Term Appointment (FTA)
        - Loan
        - Secondment
        - Permanent

    WorkLocation:
      type: string
      description: Work location
      enum:
        - Home based
        - Office based

    WorkingPattern:
      type: string
      description: Working pattern
      enum:
        - Full-time
        - Part-time

    Grade:
      type: string
      description: Grade / seniority
      enum:
        - Senior Civil Service - Permanent Secretary
        - Senior Civil Service - Director General (PB3)
        - Senior Civil Service - Director (PB2)
        - Senior Civil Service - Deputy Director (PB1/1A)
        - Grade 6
        - Grade 7
        - Senior Executive Officer (SEO)
        - Higher Executive Officer (HEO)
        - Executive Officer (EO)
        - Administrative Officer (AO)
        - Administrative Assistant (AA)

    Profession:
      type: string
      description: Profession / occupational group
      enum:
        - Actuary
        - Commercial
        - Communications
        - Corporate Finance
        - Counter Fraud
        - Digital and Data
        - Economics
        - Finance
        - Geography
        - Human Resources
        - Intelligence Analysis
        - Internal Audit
        - International Trade
        - Knowledge and Information Management
        - Legal
        - Clinical
        - Occupational Psychology
        - Operational Delivery
        - Operational Research
        - Planning
        - Planning Inspection
        - Policy
        - Project Delivery
        - Property
        - Risk Management
        - Science and Engineering
        - Security
        - Social Research
        - Statistics
        - Tax
        - Veterinary

    DCStatus:
      type: string
      description: Disability confident status
      enum:
        - Committed
        - Disability Confident Employer
        - Disability Confident Leader